backend/requirements.txt
fastapi
uvicorn
requests
PyGithub>=1.58
